Scene node, manager, and renderer classes
A basic skeleton for scene nodes and scene managers are implemented, and
is renderable by the renderer. Sleepwalker engine contains the 'engine',
while Sleepwalker itself contains the game. Huge thanks to
http://goo.gl/4FJr9 for the ideas behind this structure.

An interesting feature of the scene manager is that you can query scene
nodes by name using LINQ. An example is shown in Game1.cs in Draw().
Uncomment the second draw call (and comment the first) to try it out.